refactor(gateway): split catalog terminal start handler
